Fix Test ERROR: logs/testing.log: No such file or directory
This commit fixes the error while testing the app. The problem
is that the directory path of the shell script is invalid, even
though the dir exists. The full path to the dir must be constructed.

Adding documentation to the README.md for people coming without
python knowledge. Always good to have how-to entirely. Printing the
current time and the IP address of the caller when testing.

*	modified:   README.md
- Adding more documentation

*	modified:   dockerpuller/app.py
- Passing the ip address as the parameter to all webhooks

*	modified:   dockerpuller/scripts/testing.sh
- Computing the current dir
- Adding function to print the current timestamp
- Printing the date and any parameter passed to the log file